Moved template out of a data file and into the code, this template will slowly be eroted as the SVG lib is built up.  Separated the svgobjs lib into the more appropriate solaobjs and svgobjs libs.  Also added CLI parameters to constrain which frames are included.
